feat: add core checkout validation schemas (PR-88c-1)
Add 5 comprehensive Zod schemas for Greek marketplace checkout:
- CartLineSchema: Product cart validation with Greek error messages
- ShippingMethodSchema: Delivery options with postal code filtering
- PaymentMethodSchema: Greek market payment types (card, bank, cash)
- OrderSummarySchema: Complete order validation
- CheckoutFormSchema: Full checkout form with Greek postal/phone validation

Includes basic validation functions and Greek currency formatting.

AC: ✅ type-check PASS, ✅ build PASS, no breaking changes
LOC: 249/300 ✅
