--- frontend/tests/e2e/checkout-schema-validation.spec.ts
@@ -1,38 +1,0 @@
-/**
- * E2E Smoke Test: Checkout Schema Validation 
- * Minimal test to verify new Zod validation schemas work in real app context
- * ≤30 LOC as requested
- */
-
-import { test, expect } from '@playwright/test';
-
-test.describe('Checkout Schema Validation Smoke Test', () => {
-  test('validates checkout schemas are properly integrated', async ({ page }) => {
-    // Navigate to homepage and check if app loads
-    await page.goto('/');
-    await expect(page).toHaveTitle(/Dixis/);
-    
-    // Test schema validation by checking if validation modules can be imported
-    // This validates that our new schemas don't break the app build
-    const validationCheck = await page.evaluate(() => {
-      // Simple integration test - check if validation functions exist
-      return new Promise((resolve) => {
-        try {
-          // Try to access validation through window if exposed, or just check app didn't crash
-          const hasValidationImports = document.querySelector('[data-testid="product-card"]') !== null;
-          resolve(hasValidationImports);
-        } catch (error) {
-          resolve(false);
-        }
-      });
-    });
-    
-    expect(validationCheck).toBe(true);
-    
-    // Quick smoke test: navigate to cart (tests our new API client doesn't break routing)
-    await page.goto('/cart');
-    await expect(page.locator('body')).toBeVisible();
-    
-    console.log('✅ Checkout schema validation smoke test passed - no breaking changes detected');
-  });
-});
